fix a bug where division of ints will not return a double result

+++ src/main/java/com/mycompany/netreali/ArticleCompareHandler.java
@@ -40,7 +40,7 @@ public class ArticleCompareHandler {
             }
         }       
         
-        return foundCount / target.size();
+        return (double)foundCount / target.size();
     }
 
     public double calcSimilarity(Map<String, ArrayList<String>> target, Map<String, ArrayList<String>> source) {
